Modified game map, tiles and boundaries
Made improvements to boundaries but still needs tweaking to account for
the empty space inside the actual sprite.
changed the rocks for actual wall sprites.  modified the game board to
make it easier to see by moving the center trees and adding walls
instead.
Changed tile handling to use Switch statement because IF statements were
not as flexible and forced the used of limited logic.
Added snapshot of screen with grid for documentation in README.
Assigned values to canvas width and height in HTML for consistency until
I can make it responsive.